Move test cases to util pkg.
